Add esbuild configuration
